fonts: assume emoji presentation based on name
wezterm makes two passes over fonts when resolving emoji, based on
the presentation of the text.  If the text has a defined presentation
via a variation selector, then the first pass will only consider
fonts that match that presentation.

If no fonts match, a second pass is made to locate any font with
the appropriate glyphs.

Previously, the font was assumed to have emoji presentation if it
had color glyphs.  That meant that if you chose a monochrome emoji
font then it would be skipped for regular emoji presentation and
we'd end up matching against the default fallback for the Noto Color
Emoji font.

This commit changes the behavior to consider any explicitly listed
font with "Emoji" in the name as having emoji presentation.

This is also an imperfect heuristic, but perhaps it is good enough?

refs: #1959