fonts: remove no_glyphs cache from FontConfigInner
This avoids a confusing situation where we'd not bother to resolve
a font for a given codepoint for say bold text if we had previously
resolved it for normal text.

This hashset would be more useful if we maintained a mapping of
codepoint to font at that layer, but we don't.

The simplest thing to do for this is to allow the search to continue
and succeed.

refs: https://github.com/wez/wezterm/issues/2158
refs: https://github.com/wez/wezterm/issues/1913