ci: revise rpm filenames
I noticed that the opensuse rpms didn't show up in the latest release,
and traced it to the spec file not setting anything in the filename
to distinguish the rpm files.

This commit replaces the redhat-ish short suffix with one of our
own derived from the os info.

subst-release-info.py will need to be updated to reflect this
change, but I want to see what names are produced for the various
platforms first.