allow action=wezterm.action.ExtendSelectionToMouseCursor(nil)
In #2177 we found that one of the examples that tries to use
the default initialization was failing. The root cause is that
we were constructing a value like:

`{ExtendSelectionToMouseCursor=nil}`

which is effectively the same as:

`{}`

but that has no information on the enum variant name.

This commit avoids mapping DynValue::Null to LuaValue::Nil
and instead uses a light user data nullptr value.

That allows the structure to round trip correctly.

refs: #2177