properly fix action=wezterm.action.ExtendSelectionToMouseCursor(nil)
ee5206db507f257776230fcc3c461734802b3a5a attempted to fix this by
replacing all DynValue::Null values with a special null userdata.

However, a consequence of that is that it broke:

```
window:get_config_overrides() or {}
```

because a userdata is always true, even if it represents a null value.

This commit fixes that case by being selective with the userdata usage:
we only use it when converting the value portion of a table key/value
pair.

refs: #2177
refs: #2200