win32: associate leading dead keys with a raw key event
This lets both keys be encoded with the win32 keyboard encoding if an
invalid dead key combination is pressed. Before this change, the first
key was always sent with normal terminal encoding.

This change also works around a race condition that would frequently
occur when using win32 encoding where the first key of an invalid
combination would appear after the second.

Furthermore, pressing any two dead keys will now immediately emit both
of them. This is the same behaviour as in conhost.