fonts: allow to build gui without vendored (built-in) fonts
Bundling fonts into the binary is unnecessary when WezTerm is installed
using a package manager, e.g. from a Linux distribution repository. The
package maintainer can declare the needed fonts as dependencies of the
package, so they can be installed with the WezTerm package into a
standard location, available for all applications.

This commit adds a new feature gate "vendored-fonts" into wezterm-font
and wezterm-gui. If this feature is disabled, only
LastResortHE-Regular.ttf is included in the binary (according to [1],
WezTerm would panic if this font is not found). WezTerm will simply
use the system's font API (fontconfig on Linux) to locate the default
fonts.

This feature is enabled by default in wezterm-gui, so nothing will
change, unless explicitly disabled.

[1]: https://github.com/wez/wezterm/pull/2305#issuecomment-1194779072